Revert changes to _E_ArrayList API
While the intentions were probably good, this change was not finished,
i.e. produced an inconsistent state of the library.

Javadoc of the implemented _E_List still stated that clear() would reset
the capacity to the default (which it did not after the change). Also
other projects may rely on the existance of clear(), clear(int), reset()
and resetQuick() and we'd rather not change the API lightheartedly.

Revert "Change behavior of the clear() method to reset() used to do (reset data without reallocating the backing arrays) and remove the reset() and resetQuick() methods. Added clearQuick() method to do what resetQuick() did."

This reverts commit baa7ffccdbf165126cb2cf9c8f034b8f38657090.